Android 17.0.0 release 1

* tag 'android-17.0.0_r1': (15196 commits)
  Fix ActiveServicesTest compile failure by adding back import files
  Make ContextLogger do all of its work on the executor
  Fix privileged permission allowlist lookup in PackageManagerShellCommand
  Grant ACCESS_LOCAL_NETWORK and INTERNET permissions to AID_RADIO.
  Grant ACCESS_LOCAL_NETWORK to AID_MEDIA
  MemoryLimiter must ignore EINTR poll results
  Update appwidgetservice to not skip re-loading widgets if intent receivers were not fetched the first time due to error from package manager
  Add additional logging for when widgets are loaded.
  Fix shell access permissions for allowlists.
  Telephony: Support carrier override for network busy message
  Resolve PCC uid in ServiceRecord#postNotification and cancelNotification to avoid FGS failure
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  ...

+++ packages/SystemUI/src/com/android/systemui/clock/domain/interactor/ClockInteractor.kt
@@ -24,9 +24,11 @@ import android.icu.text.DisplayContext
 import android.os.UserHandle
 import android.provider.AlarmClock
 import androidx.annotation.VisibleForTesting
+import com.android.systemui.Flags
 import com.android.systemui.broadcast.BroadcastDispatcher
 import com.android.systemui.clock.data.repository.ClockRepository
 import com.android.systemui.dagger.SysUISingleton
+import com.android.systemui.dagger.qualifiers.Application
 import com.android.systemui.dagger.qualifiers.Background
 import com.android.systemui.plugins.ActivityStarter
 import com.android.systemui.res.R
@@ -43,6 +45,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.flatMapLatest
@@ -61,31 +64,42 @@ constructor(
     private val activityStarter: ActivityStarter,
     private val broadcastDispatcher: BroadcastDispatcher,
     private val systemClock: SystemClock,
-    @Background private val coroutineScope: CoroutineScope,
+    @Application private val applicationScope: CoroutineScope,
+    @Background private val backgroundScope: CoroutineScope,
     private val tunerService: TunerService,
 ) {
-    /** [Flow] that emits `Unit` whenever the timezone or locale has changed. */
-    val onTimezoneOrLocaleChanged: Flow<Unit> =
-        broadcastFlowForActions(Intent.ACTION_TIMEZONE_CHANGED, Intent.ACTION_LOCALE_CHANGED)
+    /** [Flow] that emits `Unit` whenever the time settings have changed. */
+    val onTimeFormatChange: Flow<Unit> =
+        broadcastFlowForActions(
+                Intent.ACTION_TIMEZONE_CHANGED,
+                Intent.ACTION_LOCALE_CHANGED,
+                Intent.ACTION_TIME_CHANGED,
+                Intent.ACTION_CONFIGURATION_CHANGED,
+                Intent.ACTION_USER_SWITCHED,
+            )
             .emitOnStart()
 
     /** [StateFlow] that emits whether the clock should show seconds. */
     val showSeconds: StateFlow<Boolean> =
-        conflatedCallbackFlow {
-                val tunable =
-                    TunerService.Tunable { key, newValue ->
-                        if (key == CLOCK_SECONDS_TUNER_KEY) {
-                            trySend(TunerService.parseIntegerSwitch(newValue, false))
+        if (!Flags.clockModernization()) {
+            MutableStateFlow(false)
+        } else {
+            conflatedCallbackFlow {
+                    val tunable =
+                        TunerService.Tunable { key, newValue ->
+                            if (key == CLOCK_SECONDS_TUNER_KEY) {
+                                trySend(TunerService.parseIntegerSwitch(newValue, false))
+                            }
                         }
-                    }
-                tunerService.addTunable(tunable, CLOCK_SECONDS_TUNER_KEY)
-                awaitClose { tunerService.removeTunable(tunable) }
-            }
-            .stateIn(
-                scope = coroutineScope,
-                started = SharingStarted.WhileSubscribed(),
-                initialValue = false,
-            )
+                    tunerService.addTunable(tunable, CLOCK_SECONDS_TUNER_KEY)
+                    awaitClose { tunerService.removeTunable(tunable) }
+                }
+                .stateIn(
+                    scope = applicationScope,
+                    started = SharingStarted.WhileSubscribed(),
+                    initialValue = false,
+                )
+        }
 
     /**
      * [StateFlow] that emits the current `Date`.
@@ -124,7 +138,7 @@ constructor(
             }
             .map { Date(systemClock.currentTimeMillis()) }
             .stateIn(
-                scope = coroutineScope,
+                scope = backgroundScope,
                 started = SharingStarted.Eagerly,
                 initialValue = Date(systemClock.currentTimeMillis()),
             )
@@ -134,11 +148,11 @@ constructor(
 
     @OptIn(ExperimentalCoroutinesApi::class)
     val longerDateFormat: Flow<DateFormat> =
-        onTimezoneOrLocaleChanged.mapLatest { getFormatFromPattern(longerPattern) }
+        onTimeFormatChange.mapLatest { getFormatFromPattern(longerPattern) }
 
     @OptIn(ExperimentalCoroutinesApi::class)
     val shorterDateFormat: Flow<DateFormat> =
-        onTimezoneOrLocaleChanged.mapLatest { getFormatFromPattern(shorterPattern) }
+        onTimeFormatChange.mapLatest { getFormatFromPattern(shorterPattern) }
 
     /** Launch the clock activity. */
     fun launchClockActivity() {
